refactor(dotcom): remove groups_frontend fork from workspaces UI
Now that workspaces has shipped, the sidebar always renders the
workspaces layout, so the groups_frontend client-side forks are dead
branches. Collapse them:

- TlaSidebar: render the workspace switcher, actions, dividers and
  recent-files list unconditionally; drop the data-workspaces attribute
- merge TlaSidebarRecentFilesNew into TlaSidebarRecentFiles and delete
  the old date-grouped, home-workspace-only version
- TlaFileMenu: gate the "Move to" submenu on admin rights alone
- TlaSidebarFileLink: enable dragging whenever the pointer is fine
- flatten the [data-workspaces='true'] CSS overrides into their base
  rules, removing the now-redundant specificity tier
- delete the now-unused useHasFlag hook
