--- apps/dotcom/client/src/tla/hooks/useHasFlag.tsx
@@ -1,11 +1,0 @@
-import { TlaFlags } from '@tldraw/dotcom-shared'
-import { useValue } from 'tldraw'
-import { useApp } from '../hooks/useAppState'
-
-export function useHasFlag(...flags: TlaFlags[]) {
-	const app = useApp()
-	return useValue('hasFlag:' + flags.join(','), () => flags.every((flag) => app.hasFlag(flag)), [
-		app,
-		flags.join(','),
-	])
-}
